pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skilllink"
version = "0.1.0"
description = "AI-powered skill and agent symlink manager for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/shaungehring/skilllink"
Repository = "https://github.com/shaungehring/skilllink"

[project.scripts]
skilllink = "skilllink.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["skilllink*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
